TheLoom/the-loom/pyproject.toml
[tool.poetry]
name = "the-loom"
version = "0.2.0"
description = "The Loom - Model server exposing hidden states for AI interpretability research (part of the Weaver ecosystem)"
readme = "README.md"
license = "Apache-2.0"
authors = ["Todd Bucy"]
keywords = ["machine-learning", "research", "interpretability", "hidden-states", "transformers", "weaver", "loom"]
packages = [{include = "src"}]

[tool.poetry.scripts]
loom = "src.server:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
torch = "^2.0.0"
transformers = "^4.57.3"
sentence-transformers = "^2.2.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pyyaml = "^6.0"
numpy = "^1.24.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
httpx = "^0.25.0"
prometheus-client = {version = "^0.19.0", optional = true}
bitsandbytes = {version = "^0.41.0", optional = true}
accelerate = "^1.12.0"
kernels = "^0.11.4"
mistral-common = {version = ">=1.8.6", python = ">=3.10,<3.14"}
triton = ">=3.4.0"
protobuf = ">=6.31.1"
sentencepiece = ">=0.1.99"

[tool.poetry.extras]
metrics = ["prometheus-client"]
quantization = ["bitsandbytes"]
all = ["prometheus-client", "bitsandbytes"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
mypy = "^1.0.0"
ruff = "^0.1.0"
types-pyyaml = "^6.0.12.20250915"
matplotlib = "^3.10.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring GPU/models",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
